requirements.txt
aiohttp
aiohttp_jinja2
jinja2
asyncio
argparse
pyyaml